Move files to UI.Inventory namespace
