+++ UI/Inventory/Hotbar.cs
@@ -1,19 +1,18 @@
 using Godot;
-using SupaLidlGame.Items;
 
-namespace SupaLidlGame.UI;
+namespace SupaLidlGame.UI.Inventory;
 
 public partial class Hotbar : GridContainer
 {
     [Export]
-    private Godot.Collections.Array<InventorySlot> _slots;
+    private Godot.Collections.Array<HotbarSlot> _slots;
 
     public override void _Ready()
     {
         Events.EventBus.Instance.PlayerInventoryUpdate += OnInventoryUpdate;
     }
 
-    public void OnInventoryUpdate(Inventory inventory)
+    public void OnInventoryUpdate(Items.Inventory inventory)
     {
         GD.Print($"UPDATE: {inventory.SelectedIndex} is selected index.");
         for (int i = 0; i < 3; i++)
