+++ UI/Inventory/ShopItem.tscn
@@ -1,0 +1,44 @@
+[gd_scene load_steps=4 format=3 uid="uid://bsheehtfcdwhh"]
+
+[ext_resource type="Theme" uid="uid://cksjbu3vrup5" path="res://UI/Themes/supalidl.tres" id="1_elbte"]
+[ext_resource type="Texture2D" uid="uid://dp7osg05ip5oo" path="res://Assets/Sprites/sword.png" id="2_5jpi0"]
+
+[sub_resource type="StyleBoxFlat" id="StyleBoxFlat_idehr"]
+bg_color = Color(0.976471, 0.956863, 0.956863, 1)
+
+[node name="ShopItem" type="BoxContainer"]
+offset_right = 67.0
+offset_bottom = 16.0
+theme = ExtResource("1_elbte")
+
+[node name="VBoxContainer" type="VBoxContainer" parent="."]
+layout_mode = 2
+theme_override_constants/separation = 0
+
+[node name="Item Margin" type="MarginContainer" parent="VBoxContainer"]
+layout_mode = 2
+
+[node name="HBoxContainer" type="HBoxContainer" parent="VBoxContainer/Item Margin"]
+layout_mode = 2
+
+[node name="TextureRect" type="TextureRect" parent="VBoxContainer/Item Margin/HBoxContainer"]
+layout_mode = 2
+texture = ExtResource("2_5jpi0")
+
+[node name="Label" type="Label" parent="VBoxContainer/Item Margin/HBoxContainer"]
+layout_mode = 2
+text = "Your mom's Item"
+
+[node name="PanelContainer2" type="PanelContainer" parent="VBoxContainer"]
+layout_mode = 2
+theme_override_styles/panel = SubResource("StyleBoxFlat_idehr")
+
+[node name="DescriptionMargin" type="MarginContainer" parent="VBoxContainer/PanelContainer2"]
+layout_mode = 2
+theme_override_constants/margin_top = 2
+theme_override_constants/margin_bottom = 2
+
+[node name="Label" type="Label" parent="VBoxContainer/PanelContainer2/DescriptionMargin"]
+layout_mode = 2
+theme_override_colors/font_color = Color(0, 0, 0, 1)
+text = "250 Shillings"
